Refactor chat types and imports across client and shared types
