+++ client/src/types/chat.ts
@@ -60,97 +60,8 @@ export interface ChatUser {
 }
 
 // 🔥 UNIFIED ChatMessage interface - مصدر واحد للحقيقة
-export interface ChatMessage {
-  id: number;
-  content: string;
-  senderId: number;
-  timestamp: string; // ✅ توحيد النوع ليكون string دائماً
-  messageType: 'text' | 'image' | 'system';
-  sender: ChatUser; // ✅ جعل المرسل مطلوب
-  
-  // Optional properties
-  roomId?: string;
-  isPrivate?: boolean;
-  receiverId?: number;
-  receiver?: ChatUser;
-}
-
-export interface PrivateConversation {
-  [userId: number]: ChatMessage[];
-}
-
-// 🔥 SIMPLIFIED WebSocket message types - حذف التعقيدات
-export type WebSocketMessageType = 
-  // Core message types
-  | 'publicMessage' | 'privateMessage' | 'newMessage'
-  // User events
-  | 'userJoined' | 'userLeft' | 'onlineUsers' | 'typing'
-  // Room events  
-  | 'roomJoined' | 'roomMessages'
-  // System events
-  | 'auth' | 'error' | 'warning'
-  // Misc events
-  | 'kicked' | 'profileUpdated';
-
-export type WebSocketMessage = {
-  type: WebSocketMessageType;
-  
-  // خصائص الرسالة الأساسية
-  username?: string;
-  content?: string;
-  messageType?: string;
-  receiverId?: number;
-  isTyping?: boolean;
-  user?: ChatUser;
-  users?: ChatUser[];
-  message?: ChatMessage | string; // يمكن أن يكون string أيضاً
-  messages?: ChatMessage[]; // للرسائل المجمّعة مثل roomMessages
-  broadcastInfo?: any; // معلومات غرفة البث المرفقة مع الحدث
-  action?: string;
-  
-  // خصائص إضافية للوظائف المختلفة
-  isHidden?: boolean; // لرؤية المستخدم
-  color?: string; // لتغيير لون اسم المستخدم
-  usernameColor?: string; // لون اسم المستخدم الجديد
-  userTheme?: string; // لموضوع المستخدم
-  profileEffect?: string; // لتأثير البروفايل
-  targetUserId?: number; // للإجراءات المستهدفة
-  duration?: number; // مدة الإجراء (كيك، حظر، إلخ)
-  reason?: string; // سبب الإجراء
-  moderatorName?: string; // اسم المراقب
-  notificationType?: string; // نوع الإشعار
-  senderId?: number; // معرف المرسل
-  senderUsername?: string; // اسم المرسل
-  acceptedBy?: string; // من قبل الطلب
-  friendId?: number; // معرف الصديق
-  newRole?: string; // الدور الجديد للترقية
-  
-  // خصائص إضافية مفقودة للتوافق مع الاستخدامات
-  userId?: number; // إضافة: معرف المستخدم
-  source?: string; // إضافة: مصدر الرسالة لأغراض السجلات
-  
-  // خصائص النقاط
-  points?: number; // عدد النقاط
-  senderName?: string; // اسم مرسل النقاط
-  receiverName?: string; // اسم مستقبل النقاط
-  oldLevel?: number; // المستوى القديم
-  newLevel?: number; // المستوى الجديد
-  levelInfo?: any; // معلومات المستوى
-  
-  // خصائص Broadcast Room الجديدة
-  roomId?: string; // معرف الغرفة
-  hostId?: number; // معرف المضيف
-  speakers?: number[]; // قائمة المتحدثين
-  micQueue?: number[]; // قائمة انتظار طلبات المايك
-  requestUserId?: number; // معرف المستخدم الذي طلب المايك
-  speakerId?: number; // معرف المتحدث
-  speakerAction?: 'add' | 'remove'; // نوع عملية المتحدث
-  queueAction?: 'join' | 'leave' | 'accept' | 'reject'; // نوع عملية طابور المايك
-  
-  // خصائص Wall Post
-  post?: any; // بيانات المنشور
-  postId?: number; // معرف المنشور
-};
+export type { ChatMessage } from '../../../shared/types';
+import type { ChatMessage as SharedChatMessage } from '../../../shared/types';
 
 export interface UserProfile {
   name: string;
@@ -231,11 +142,12 @@ export interface RoomUser extends ChatUser {
   joinedAt: Date;
 }
 
-export interface RoomMessage extends ChatMessage {
+export interface RoomMessage extends SharedChatMessage {
   roomId: string;
 }
 
-export interface RoomWebSocketMessage extends WebSocketMessage {
+import type { WebSocketMessage as SharedWSMessage } from '../../../shared/types';
+export interface RoomWebSocketMessage extends SharedWSMessage {
   roomId?: string;
   room?: ChatRoom;
   rooms?: ChatRoom[];
