--- server/db.ts
@@ -1,14 +1,0 @@
-// استيراد المحول الجديد
-export { 
-  db, 
-  dbType, 
-  dbAdapter, 
-  checkDatabaseHealth, 
-  getDatabaseStatus,
-  type DatabaseType,
-  type DatabaseAdapter
-} from './database-adapter';
-
-// للتوافق مع الكود الموجود
-export const pool = null; // سيتم إزالته تدريجياً
-export const sqlite = null; // سيتم إزالته تدريجياً
